requirements.txt
###### Requirements without Version Specified ######`
PyQt5
pyowm
qtawesome
qt-material
geocoder
requests
pytz
yelpapi
matplotlib
pillow
